Revise the 'reducedRatio' infrastructure.
This fix is mostly cosmetic, but it also includes one notable functional change:
The purpose of function `reducedRatio` is to reduce the components of a given ratio to be <= a given threshold.
Therefore, it does not need a special case of reducing that ratio even further to `1 / 1`.
If we wanted to reduce it further, as much as possible, then the full solution would require applying GCD.
Reducing to `1 / 1` is just a (probably rare) special-case handling, which does not serve the original purpose.

This functional change is applied in two places:
1. In function `reducedRatio`, the `1 / 1` return-statement is removed
2. In function `estimatedRatio`, the last return-statement is slightly modified

Cosmetic changes include:
1. Improve documentation in the related functions
2. Improve variable naming (replace `Fraction memory r` with `Fraction memory ratio`)
3. Remove the redundant usage of `newR` and `newRatio` (input parameter `ratio` is sufficient)