Simplify the 'reducedRatio' infrastructure.
Previous implementation:
Try to ensure that both components remain as large as possible (i.e., the larger component remains as close as possible to the given threshold).

New implementation:
Just ensure that both components do not exceed the given threshold (i.e., the larger component is not necessarily as close as possible to the given threshold).

For example, given the ratio `10 / 13` and the threshold `12`:
- The previous implementation returns the ratio `5 / 7`, which represents the value `0.71428`
- The new implementation returns the ratio `5 / 6`, which represents the value `0.83333`
- The input ratio `10 / 13` represents the value `0.76923`

Note that as expected, both implementations reduce the components of the given ratio to the given threshold.
However, the input ratio is closer to the output of the previous implementation than to the output of the new implementation.

For a large range of values, each implementation "wins" some cases and "loses" some cases.
However, as you can see, the new implementation is much simpler and cleaner than the previous implementation.