**Contract `AutoCompoundingStakingRewards`:**
In the current implementation of the Flat Program, the first user is subjected to some precision-loss.
Then, the remaining rewards are distributed to the remaining users.
So whatever the first user loses, is eventually distributed among the remaining users.
Same goes for the second user, and so forth...
Regardless of the above, the precision-loss is negligible.
Nevertheless, we may as well change the implementation such that all users are subjected to the exact same precision-loss.

**Contract `StakingRewardsMath`:**

1. Change the Flat function to take "total" values instead of "remaining" values.
2. Change the Flat function input order to: `totalRewards`, `timeElapsed`, `programDuration`.
3. Change the Exp-Decay function input order to: `totalRewards`, `timeElapsed`.
4. In the Flat function, cap the value of `timeElapsed` to at most the value of `programDuration` (instead of doing it in contract `AutoCompoundingStakingRewards`).