Optimize gas consumption in contract `PendingWithdrawals`:
1. Function `isReadyForWithdrawal` reads at most 2 slots from the storage, hence no point in reading an entire `WithdrawalRequest` entry into memory.
Note that although this function is read-only, this change saves ~6.6k gas when called from a non-read-only function.
Note that it can be reduced by ~2.2k more gas, by rearranging  `struct WithdrawalRequest`, such that `address provider` and `uint32 createdAt` are located in the same slot.

2. Function `reinitWithdrawal` reads all the slots of a `WithdrawalRequest` entry, so it makes sense to read this entire entry into memory, instead of referencing it from storage.
However, doing so seems to add ~2k gas for some reason, so no change applied for now.

3. Function `_removeWithdrawalRequest` receives an entire `WithdrawalRequest` entry as input, but it requires only the `provider` field.
This function is called only from functions which already read the entire entry into memory, so there isn't much to gain by passing only the required field at that point (only 6 gas).
However, on top of the fact that it is probably a better practice to pass only the required input, it might also reveal other storage/memory potential improvements, hence this change is applied.